Sprite lighting in + Zindexing
Sprites and batches can now have an zindex.

Lighting is now handled by frame buffers.

First we render the scene to a frame buffer.
Second render sprite lights in frame buffer;
Third combine the first two frame buffers and multiply them in the shader. This is all done in the game renderer.